+++ src/Services/Feed/Feed.Domain/Aggregates/Author/AuthorPermission.cs
@@ -4,14 +4,14 @@ namespace Feed.Domain.Aggregates.Author {
     public class AuthorPermission : Entity {
         public long UserId { get; private set; }
         public PermissionScope Scope { get; private set; }
-        public int Flags { get; private set; }
+        public short Flags { get; private set; }
 
-        public AuthorPermission(PermissionScope scope, int flags) {
+        public AuthorPermission(PermissionScope scope, short flags) {
             Scope = scope;
             Flags = flags;
         }
 
-        internal void AddFlags(int flags) {
+        internal void AddFlags(short flags) {
             Flags |= flags;
         }
     }
